fix(openui-cloud): make the example build in the workspace
The openui-cloud example was excluded from the workspace build
(`!examples/openui-cloud` in pnpm-workspace.yaml) and did not build. Re-enable
it and fix the build:

- Re-add openui-cloud to the pnpm workspace and refresh the lockfile.
- Bump next / eslint-config-next to 16.2.6 to match the other examples.
- Split the single client `page.tsx` into a server page that renders a
  `"use client"` CloudChat via `export const dynamic = "force-dynamic"`. This
  stops Turbopack from statically prerendering the heavy artifact-renderer SSR
  chunk at build time (it emitted an invalid chunk: "Identifier 'h' has already
  been declared"), matching the CLI template structure.
