+++ examples/openui-cloud/package.json
@@ -30,7 +30,7 @@
     "lucide-react": "^0.575.0",
     "mdast-util-find-and-replace": "3.0.2",
     "mermaid": "11.15.0",
-    "next": "16.1.6",
+    "next": "16.2.6",
     "react": "19.2.3",
     "react-dom": "19.2.3",
     "recharts": "2.15.4",
@@ -49,7 +49,7 @@
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
-    "eslint-config-next": "16.1.6",
+    "eslint-config-next": "16.2.6",
     "openai": "^6.22.0",
     "tailwindcss": "^4",
     "typescript": "^5",
