+++ pnpm-workspace.yaml
@@ -3,7 +3,6 @@ packages:
   - "examples/**/*"
   - "docs/"
   - "!**/src/templates/**"
-  - "!examples/openui-cloud"
 
 # Centralized dependency versions shared across packages.
 # Reference these from a package.json with the "catalog:" protocol.
